+++ lib/ecto/datetime.ex
@@ -3,27 +3,39 @@ defmodule Ecto.Date do
   An Ecto type for dates.
   """
 
+  @behaviour Ecto.Type
   defstruct [:year, :month, :day]
+
+  @doc """
+  The Ecto primitive type.
+  """
+  def type, do: :date
+
+  @doc """
+  Casts to date.
+  """
+  def cast(%Ecto.Date{} = d), do: {:ok, d}
+  def cast(_), do: :error
 
   @doc """
   Converts an `Ecto.Date` into a date triplet.
   """
-  def to_erl(%Ecto.Date{year: year, month: month, day: day}) do
-    {year, month, day}
+  def dump(%Ecto.Date{year: year, month: month, day: day}) do
+    {:ok, {year, month, day}}
   end
 
   @doc """
   Converts a date triplet into an `Ecto.Date`.
   """
-  def from_erl({year, month, day}) do
-    %Ecto.Date{year: year, month: month, day: day}
+  def load({year, month, day}) do
+    {:ok, %Ecto.Date{year: year, month: month, day: day}}
   end
 
   @doc """
   Returns an `Ecto.Date` in local time.
   """
   def local do
-    from_erl(:erlang.date)
+    load(:erlang.date) |> elem(1)
   end
 
   @doc """
@@ -31,7 +43,7 @@ defmodule Ecto.Date do
   """
   def utc do
     {date, _time} = :erlang.universaltime
-    from_erl(date)
+    load(date) |> elem(1)
   end
 end
 
@@ -40,27 +52,39 @@ defmodule Ecto.Time do
   An Ecto type for time.
   """
 
+  @behaviour Ecto.Type
   defstruct [:hour, :min, :sec]
+
+  @doc """
+  The Ecto primitive type.
+  """
+  def type, do: :time
+
+  @doc """
+  Casts to time.
+  """
+  def cast(%Ecto.Time{} = t), do: {:ok, t}
+  def cast(_), do: :error
 
   @doc """
   Converts an `Ecto.Time` into a time triplet.
   """
-  def to_erl(%Ecto.Time{hour: hour, min: min, sec: sec}) do
-    {hour, min, sec}
+  def dump(%Ecto.Time{hour: hour, min: min, sec: sec}) do
+    {:ok, {hour, min, sec}}
   end
 
   @doc """
   Converts a time triplet into an `Ecto.Time`.
   """
-  def from_erl({hour, min, sec}) do
-    %Ecto.Time{hour: hour, min: min, sec: sec}
+  def load({hour, min, sec}) do
+    {:ok, %Ecto.Time{hour: hour, min: min, sec: sec}}
   end
 
   @doc """
   Returns an `Ecto.Time` in local time.
   """
   def local do
-    from_erl(:erlang.time)
+    load(:erlang.time) |> elem(1)
   end
 
   @doc """
@@ -68,7 +92,7 @@ defmodule Ecto.Time do
   """
   def utc do
     {_date, time} = :erlang.universaltime
-    from_erl(time)
+    load(time) |> elem(1)
   end
 end
 
@@ -77,6 +101,7 @@ defmodule Ecto.DateTime do
   An Ecto type for dates and times.
   """
 
+  @behaviour Ecto.Type
   defstruct [:year, :month, :day, :hour, :min, :sec]
 
   @doc """
@@ -85,18 +110,24 @@ defmodule Ecto.DateTime do
   def type, do: :datetime
 
   @doc """
+  Casts to date time.
+  """
+  def cast(%Ecto.DateTime{} = dt), do: {:ok, dt}
+  def cast(_), do: :error
+
+  @doc """
   Converts an `Ecto.DateTime` into a `{date, time}` tuple.
   """
-  def to_erl(%Ecto.DateTime{year: year, month: month, day: day, hour: hour, min: min, sec: sec}) do
-    {{year, month, day}, {hour, min, sec}}
+  def dump(%Ecto.DateTime{year: year, month: month, day: day, hour: hour, min: min, sec: sec}) do
+    {:ok, {{year, month, day}, {hour, min, sec}}}
   end
 
   @doc """
   Converts a `{date, time}` tuple into an `Ecto.DateTime`.
   """
-  def from_erl({{year, month, day}, {hour, min, sec}}) do
-    %Ecto.DateTime{year: year, month: month, day: day,
-                   hour: hour, min: min, sec: sec}
+  def load({{year, month, day}, {hour, min, sec}}) do
+    {:ok, %Ecto.DateTime{year: year, month: month, day: day,
+                         hour: hour, min: min, sec: sec}}
   end
 
   @doc """
@@ -126,13 +157,13 @@ defmodule Ecto.DateTime do
   Returns an `Ecto.DateTime` in local time.
   """
   def local do
-    from_erl(:erlang.localtime)
+    load(:erlang.localtime) |> elem(1)
   end
 
   @doc """
   Returns an `Ecto.DateTime` in UTC.
   """
   def utc do
-    from_erl(:erlang.universaltime)
+    load(:erlang.universaltime) |> elem(1)
   end
 end