+++ test/ecto/query/planner_test.exs
@@ -14,7 +14,7 @@ defmodule Ecto.Query.PlannerTest do
     schema "comments" do
       field :text, :string
       field :temp, :string, virtual: true
-      field :posted, Custom.DateTime
+      field :posted, Ecto.DateTime
       belongs_to :post, Ecto.Query.PlannerTest.Post
       has_many :post_comments, through: [:post, :comments]
     end
@@ -92,7 +92,7 @@ defmodule Ecto.Query.PlannerTest do
   end
 
   test "prepare: casts and dumps custom types" do
-    datetime = %Custom.DateTime{year: 2015, month: 1, day: 7, hour: 21, min: 18, sec: 13}
+    datetime = %Ecto.DateTime{year: 2015, month: 1, day: 7, hour: 21, min: 18, sec: 13}
     {_query, params} = prepare(Comment |> where([c], c.posted == ^datetime))
     assert params[0] == {{2015, 1, 7}, {21, 18, 13}}
 
@@ -102,7 +102,7 @@ defmodule Ecto.Query.PlannerTest do
   end
 
   test "prepare: casts and dumps custom types with arrays" do
-    datetime = %Custom.DateTime{year: 2015, month: 1, day: 7, hour: 21, min: 18, sec: 13}
+    datetime = %Ecto.DateTime{year: 2015, month: 1, day: 7, hour: 21, min: 18, sec: 13}
     {_query, params} = prepare(Comment |> where([c], c.posted in ^[datetime]))
     assert params[0] == [{{2015, 1, 7}, {21, 18, 13}}]